more abstracting and refactoring
modified the config.yml file so that most changes that need to be made
by the user can be done there. Such as bg-img, header title, logo,
blurb etc.

started trying to center the skills section a bit better. I wanted to
save it in it’s working condition so I commented out the parts I was
using to experiment with.