success: _data/portfolio-folder
successfully moved most of config.yml to _data/portfolio folder where a
lot of the user customizations can be setup. the only thing that I
could not seem to get working was the jekyll sass variables, so the
header, footer, logo, and avatar images are hard coded. It works though
:).